+++ examples/hkl_test.py
@@ -2,9 +2,8 @@ from __future__ import print_function
 
 import config
 
-from ophyd.hkl.hkl import CalcRecip
 from ophyd.hkl.diffract import E4CH
-from ophyd.hkl.calc import (CalcE4CH, CalcK6C)
+from ophyd.hkl.calc import (CalcRecip, CalcE4CH, CalcK6C)
 from ophyd.hkl.util import diffractometer_types
 import ophyd.hkl.hkl as hkl_module
 from ophyd.controls.positioner import Positioner
